% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_modifications}
\alias{calc_modifications}
\alias{calc_A_export}
\alias{calc_B_export}
\title{Using Both Subdirectory Functions}
\usage{
calc_modifications(x)

calc_A_export(x)

calc_B_export(x)
}
\arguments{
\item{x}{Vector of Numeric Values with length \eqn{n}.}
}
\value{
A vector of numeric values with length \eqn{n}.
}
\description{
These function serve as intermediaries between R and the subdirectory
source.
}
\details{
The exported function name is different from the subdirectory function. 
C++ does not allow duplicate function names with the same parameters. To 
preserve the original function name necessitates the subdirectory function
being placed behind a namespace, e.g. `A::calc_A_routine()`. Otherwise,
the name of the function exported must be different, e.g. `calc_A_export()`.
}
\examples{
my_vec = c(3.5, 8, -1.5, -2, 0, 42)
calc_modifications(my_vec)
my_vecA = c(881, 884, 1,2,-4)
calc_A_export(my_vec)
my_vecB = c(512,32, -1, 1.3, 5.9)
calc_B_export(my_vecB)
}
