% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubdirSrc-package.R
\docType{package}
\name{SubdirSrc-package}
\alias{SubdirSrc}
\alias{SubdirSrc-package}
\title{SubdirSrc: Embedding Compiled Code In Subdirectories of Src}
\description{
Provides an implementation of how to modify a Makevars file to
permit compilation of code found in a subdirectory of the `src` directory.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-headers-subdirs}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
